refactor book detials screen so one activity neded
