+++ src/app/src/main/AndroidManifest.xml
@@ -17,12 +17,9 @@
         <activity android:name=".ScanIsbnActivity" />
         <activity android:name=".RequestedBookDetailsActivity" />
         <activity android:name=".PendingRequestsActivity" />
-        <activity android:name=".MyRequestedBookDetailsActivity" />
         <activity android:name=".MyProfileActivity" />
-        <activity android:name=".MyBorrowedBookDetailsActivity" />
         <activity android:name=".MyBooksActivity" />
-        <activity android:name=".MyAvailableBookDetailsActivity" />
-        <activity android:name=".MyAcceptedBookDetailsActivity" />
+        <activity android:name=".MyBookDetailsActivity" />
         <activity android:name=".ManageRequestsActivity" />
         <activity android:name=".ExploreBookDetailsActivity" />
         <activity android:name=".ExploreActivity" />
