--- src/app/src/main/java/com/example/bookmark/MyRequestedBookDetailsActivity.java
@@ -1,18 +1,0 @@
-package com.example.bookmark;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-/**
- * TODO: Description of class.
- * @author Mitch Adam.
- */
-public class MyRequestedBookDetailsActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_requested_book_details);
-    }
-}
