+++ src/app/src/main/res/layout/activity_my_book_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyAvailableBookDetailsActivity">
+    tools:context=".MyBookDetailsActivity">
 
     <TextView
         android:id="@+id/book_details_title_text"
@@ -64,7 +64,7 @@
 
     <TextView
         android:id="@+id/book_details_book_status_text"
-        android:layout_width="190dp"
+        android:layout_width="219dp"
         android:layout_height="30dp"
         android:layout_marginTop="16dp"
         android:text="Status: Available"
@@ -72,5 +72,17 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/book_details_book_image" />
+
+    <Button
+        android:id="@+id/book_details_action_btn"
+        android:layout_width="219dp"
+        android:layout_height="38dp"
+        android:layout_marginTop="16dp"
+        android:background="#673AB7"
+        android:text="Action"
+        android:textColor="#FFFFFF"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/book_details_book_status_text" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
